Fix home page and examples pages to work with typescript output
