+++ examples/index.html
@@ -7,11 +7,17 @@
     <script src="../dist/easeljs.js"></script>
     <script type="text/javascript" src="../dist/pathseg.js"></script>
     <script type="text/javascript" src="../dist/txt.js"></script>
-    <script type="text/javascript" src="../dist/examples.js"></script>
+
   </head>
   <body
     style="padding-left: 160px; padding-top: 10px; background-color: antiquewhite;"
   >
+  <div id="layout">
+    <!-- Menu toggle -->
+    <a href="#menu" id="menuLink" class="menu-link">
+      <!-- Hamburger icon -->
+      <span></span>
+    </a>
     <div id="menu">
       <div class="pure-menu pure-menu-open">
         <a class="pure-menu-heading" href="../">txtjs</a>
@@ -30,5 +36,11 @@
         </ul>
       </div>
     </div>
+  </div>
+    <script type="text/javascript" src="../dist/site.js"></script>
+    <script type="text/javascript">
+      txt.FontLoader.path = "../font/";
+      site.initExampleLoader();
+    </script>
   </body>
 </html>
