+++ index.html
@@ -387,5 +387,12 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 </div>
 
 <script type="text/javascript" src="dist/site.js"></script>
+<script type="text/javascript">
+	txt.FontLoader.path = "font/";
+	const exampleEl = document.getElementById("example");
+	if(exampleEl){
+		site.initDemo(exampleEl);
+	}
+  </script>
 </body>
 </html>
